+++ ast/tree/INode.h
@@ -7,26 +7,6 @@
 namespace NSyntaxTree {
     class IVisitor;
 
-    struct Position {
-        uint32_t line;
-        uint32_t column;
-
-        string ToString() const {
-            std::stringstream buf;
-            buf << "("  << line << ", " << column << ")";
-            return buf.str();
-        }
-    };
-
-    struct Location {
-        Position begin;
-        Position end;
-
-        string ToString() const {
-            return begin.ToString() + " - " + end.ToString();
-        }
-    };
-
     interface INode {
         Location location;
