+++ ast/tree/common.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <util/CommonDefs.h>
+#include <util/Location.h>
 
 #include <memory>
 #include <string>
@@ -10,4 +11,7 @@ namespace NSyntaxTree {
     using std::unique_ptr;
     using std::vector;
     using std::string;
+
+    using NUtil::Location;
+    using NUtil::Position;
 }
