+++ symbol_table/common.h
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <ast/tree/INode.h>
-
+#include <util/Location.h>
 #include <util/Symbol.h>
 #include <util/StringInterner.h>
 
@@ -9,5 +8,5 @@ namespace NSymbolTable {
     using NUtil::Symbol;
     using NUtil::StringInterner;
 
-    using NSyntaxTree::Location;
+    using NUtil::Location;
 }
