+++ type_checker/Exceptions.h
@@ -2,20 +2,21 @@
 
 #include <ast/Exceptions.h>
 #include <symbol_table/TypeInfo.h>
+#include <util/Location.h>
 
 namespace NTypeChecker {
     class IllegalTypeException: public NSyntaxTree::SyntaxError {
     public:
         using NSyntaxTree::SyntaxError::SyntaxError;
 
-        IllegalTypeException(const NSyntaxTree::Location &loc
+        IllegalTypeException(const NUtil::Location &loc
             , const NSymbolTable::TypeInfo &found
             , const NSymbolTable::TypeInfo &expected)
             : SyntaxError("Error: expected " + expected.ToString() + ", found " + found.ToString()
                           + " on " + loc.ToString()) {
         }
 
-        IllegalTypeException(const NSyntaxTree::Location &loc
+        IllegalTypeException(const NUtil::Location &loc
             , const NSymbolTable::TypeInfo &found
             /* expected non POD type */)
             : SyntaxError("Error: expected object of some class, found " + found.ToString()
@@ -26,7 +27,7 @@ namespace NTypeChecker {
     class BadArgumentsException: public NSyntaxTree::SyntaxError {
     public:
         using NSyntaxTree::SyntaxError::SyntaxError;
-        explicit BadArgumentsException(const NSyntaxTree::Location &loc)
+        explicit BadArgumentsException(const NUtil::Location &loc)
             : SyntaxError("Error: bad arguments on " + loc.ToString()) {
         }
     };
@@ -34,7 +35,7 @@ namespace NTypeChecker {
     class PrivateAccessException: public NSyntaxTree::SyntaxError {
     public:
         using NSyntaxTree::SyntaxError::SyntaxError;
-        PrivateAccessException(const NSyntaxTree::Location &loc
+        PrivateAccessException(const NUtil::Location &loc
                 , const NSymbolTable::Symbol* methodId
                 , const NSymbolTable::Symbol* classId)
                 : SyntaxError("Unexpected access to private method "
